diffapply: make overwritten dirs opaque.
A previous commit fixed an issue where directories that overwrite
whiteout devices needed to be set to be opaque. While that was needed,
it wasn't enough because directories also need to be set opaque if they
are overwriting anything that isn't a directory, such as a file or a
symlink. This commit updates the code to handle that situation in
general.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>