snapshot: use temporary lease during diffapply.
A temporary lease ctx was previously made during the snapshotter Merge
method but it was not passed to diffApply, which may have meant that
temporary View snapshots created there could be garbage collected by
containerd at any time.

It's possible this was encountered in CI, but I was unable to reproduce
the failure after 1000 attempts locally, so this commit is just a best
guess at the problem.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>
(cherry picked from commit 868f78d0701adac3f4c943e2e93a9470050ebaf0)
Signed-off-by: Tonis Tiigi <tonistiigi@gmail.com>