refactor to use a wrapper MkidrTemp function
This is because docker create `TMPDIR` at`/var/lib/docker/tmp` it
only allows `root:root` user to access. However, this means
when we use userns_remap mode in docker daemon, we will not
be able to mount any volume in the container from this tmp directory

The solution is to create a separate directory under the env var
`BUILDKIT_TMPDIR`, and treat it as the default temp directory.

Signed-off-by: Jiashu Chen <cjs20080808@hotmail.com>