[CUSTOM] Stop returning errors for response retries when request is a cross repo blob mount.
The upstream pusher client from containerd for docker attempts an optimization
where instead of uploading a blob it will mount it if available from another repo. In the
event that the blob can't be read from the other repo, due to lacking authorization, then it
falls back to the original, un-optimized, upload. Throwing an error in this situation from buildkit
prevents containerd code from falling back to the original upload request resulting in a failed build/upload.

Signed-off-by: Jacob MacElroy <jacob@okteto.com>