progressui: fix printing of hidden vertexes.
The plain progress has a loop that iterates over each vertex the trace
knows about (rather than just the updated ones). Before this, that loop
was not checking to see if a vertex was supposed to be skipped due to
being in a group or due to being hidden.

This resulted in vertexes being printed unnecessarily. It also led to a
rare case where a vertex with nil lastBlockTime could be set to the
'current' field, which could then result in a nil pointer exception.

The update here just adds the appropriate checks to the loop.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>