util/resolver/authorizer.go: Add basic auth support based on existing login
The authorized coded is missing support for the following scenario:

1. docker login to private registry (e.g. harbor)
2. buildctl build --output type=image,name=privateregistry/project/repo,push=true

When the private registry is first contacted, there is no authHandler being
created with the user id and password from the .docker/config.json file. The
implementation falls back to oauth authentication which might fail especially
if locally created ca/certificates are used.

This commit adds support for this scenario in the get authHandler function.