exporter: introduce annotations support
This patch adds support for attaching annotations to the oci and
containerimage exporters.

Annotations can be created by frontends by attaching them to the result
metadata, or from the parameter keys passed to the exporter.

Annotations can be attached to 4 different levels:
- Image manifest (per platform)
- Image manifest descriptors in the image index (per platform)
- Image index
- Image index descriptors in the OCI image layout

Note that the only one the image manifest is always guaranteed to be
present (since an image index is not always created).

Signed-off-by: Justin Chadwell <me@jedevc.com>