dockerfile: restructure parser packages
This patch renames the parser to the ast package, and then creates a new
parser package to contain both the instructions and new ast package.
This allows implementing a new helper for parsing dockerfiles, that can
remain API stable regardless of underlying changes in the parsing
mechansisms used, as well as providing a more helpful user interface to
those performing programatic analysis of Dockerfiles.

Signed-off-by: Justin Chadwell <me@jedevc.com>