tests: extract registry image info helpers to testutil
The image info registry helpers provide utilities for reading content
from providers such as registries which can be easily used to inspect
image content. This is an alternative to requiring the containerd
puller, or exporting as an oci tarball.

By moving the helpers into testutil, we allow other testing components
to utilize the helpers, instead of just the dockerfile tests.

Signed-off-by: Justin Chadwell <me@jedevc.com>