solver: add new generic results package
The Result struct in this package is intended to replace the various
Result structs across the codebase, including:

- frontend.Result
- client.Result/gateway.Result
- exporter.Source

These Results are all fundamentally the same, and share the same logic,
with the only difference being the type used in the Ref/Refs properties.

Since these were all separate, we easily encounter the following
problems:

- Repetition of struct and method declarations.

  Each separate Result type required it's own properties and methods to
  be redeclared. Sometimes these were consistent with each other, and
  other times they were not. Unifying all Results together ensures a
  lack of duplication, and means that changing the structure of the Result
  will not require extensive changes throughout every single declaration.

- Complex conversion between Result types.

  At gateway and api boundaries, we are required to convert each Result
  into a Result of a different type. Previously, this has been a lengthly
  and error prone conversion, requiring knowledge of the exact structure
  of the Result, so as to allow traversal. By unifying Results, we can
  implement a ConvertResult function to convert from types of Result[A]
  to types of Result[B] given a function from A to B, which cleanly
  encapsulates the structure of the Result.

- Complex conversion to the protobuf gateway api types.

  At the gateway api, we have to convert to pb.Results. Previously, we
  had to include logic to convert to and from both frontend.Result and
  client.Result. With a generic Result, we can simple convert to a
  Result[pb.Ref] and then have one single set of methods to convert
  Result[pb.Ref] to a pb.Result.

These problems have been an issue for some time, however, recently with
the addition of attestations, the problem has grown, since the amount of
logic required at conversion boundaries has massively increased.

This patch intends to simplify all of the above, and reduce line-count
and code complexity by using the new-ish go generics features. Care has
been taken to ensure that this patch is as small as possible, and
affects as few Go APIs as possible. This is ensured by using type aliases
for the Result type to re-create the various previous Results, e.g.
frontend.Result/client.Result/exporter.Source, so that little code in
packages that utilize these Results need to be changed. The only major
changes take place at conversion boundaries, where, for the most part,
the code is simplified.

One major caveat is that the result package is made significantly more
complex, due to the generic code, and the need for reflection to perform
comparisons of any types to nil, as interfaces such solver.ResultProxy
(used as a Ref type) do not implement the comparable interface, which is
a limitation of the current go language spec (hopefully to be resolved
in the future).

Signed-off-by: Justin Chadwell <me@jedevc.com>