remove resolv.conf cleanup
cleanup isn't safe because many containers may require the same config,
and we don't want it to be removed until they're _all_ done with it.

so just let them stick around, since there probably won't be that many
over time anyway (...right?)

alternatively we could do what we do for /etc/hosts: make a new distinct
file if alternative config is specified and clean it up after. however
the range of resolv.conf values should be much more limited so I would
guess there's more value in sharing a single file across all of them
rather than creating and deleting them all the time.

keeping this a separate commit so it's easy to revert if we want to do
that anyway.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>