[0.10] Update uses of Image platform fields in OCI image-spec
The OCI image spec is considering to change the Image struct and embedding the
Platform type (see opencontainers/image-spec#959) in the go implementation.
BuildKit currently uses some struct-literals to propagate the platform fields,
which will break once those changes in the OCI spec are merged.

Ideally (once that change arrives) we would update the code to set the Platform
information as a whole, instead of assigning related fields individually, but
in some cases in the code, image platform information is only partially set
(for example, OSVersion and OSFeatures are not preserved in all cases). This
may be on purpose, so needs to be reviewed.

This patch keeps the current behavior (assigning only specific fields), but
removes the use of struct-literals to make the code compatible with the
upcoming changes in the image-spec module.

This patch is similar to f51b77934e2a3af1045a548050da57bcc25b2069, but contains
some changes from a163f58c77e571a760a85db8b12b2c8bc4563e61, which updated the
exporter `emptyImageConfig` to use the OCI image-spec, instead of an ad-hoc
type (which was no longer needed), and renamed it to `defaultImageConfig`.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>