dockerfile: allow tag based resolution for oci-layout contexts
This patch allows ommiting the digest in oci-layout named contexts, and
using a tag instead.

Unfortunately, client changes are required to expose the contents of
index.json - to do this, we wrap the containerd local store, and
override the .Walk method to process queries that ask for only the
indexed contents, instead of all the blobs.

We allow for two schemes:
- A full tag, e.g. oci-layout://<store>/alpine, which will look for the
  canonical docker.io/library/alpine:latest in index.json
- A partial tag, e.g. oci-layout://<store>:latest, which will search for
  an image in index.json that has a :latest tag.

  Interally, we support this by translating this to an image
  <store>/*:latest, so that we can transport it through the OCI image
  identifier structure.

Signed-off-by: Justin Chadwell <me@jedevc.com>