oci exporter: support unpack option
This feature adds support for specifying unpack=true to the oci exporter
options to unpack the resulting result for the client.

To do this, we setup a content store on the client, and forward it
through to the server, which can then copy the exported data into the
content store.

Signed-off-by: Justin Chadwell <me@jedevc.com>