attestations: add attestation bundles
An attestation bundle is a directory containing attestations.

It uses the same Ref and Path properites as an InToto Attestation, but
looks up the files inside the Ref+Path to process as raw InToto
statements. This allows flexibility for generating more complex
combinations of attestations, where all the details may not be known
ahead of time, and unlazying the reference is not desired.

Signed-off-by: Justin Chadwell <me@jedevc.com>