solver: add post-processor to convert Ref to Refs map
This adds a new processor to pass to Solve that converts a result's unit Ref
output to a simple Refs map type, which is required to be able to add
attestations at this point.

All metadata is appropriately converted, so that the new result should
have the same semantics as the old one.

Signed-off-by: Justin Chadwell <me@jedevc.com>