lint: enable gosec
By default we exclude potential hardcoded credentials and implicit
memory aliasing for giving lots of false positives on correct behavior
(where the addressed value never leaves the loop, or we break
immediately after).

We additionally exclude TLS MinVersion warnings (we need to co-ordinate
with other projects upstream if we want to do this).

Finally, we exclude net/http/cgi blocklist, since it error for a
vulnerability in a version of Go that we don't target.

Signed-off-by: Justin Chadwell <me@jedevc.com>