util: fixup staticfs sorting order
The character "." (and others) appear before "/" in the ASCII table so
when sorting, these will appear directly after the directory, before the
contents.

When the results of the sort are then used for calls to Walk(), the
traversal order will be incorrect, causing fsutil's filesync
functionality to reject the server's transfer.

To fix this, we replace "/" in the pathname with the null character, to
ensure that it is sorted before the others.

Signed-off-by: Justin Chadwell <me@jedevc.com>