dockerfile: allow new formats for syntax directives
This patch introduces support for allowing multiple formats for syntax
directives, to determine the syntax of the file to forward to the
appropriate gateway. Additionally, support for shebangs for syntax
detection is also permitted.

This does not change the dockerfile syntax and allow different
directives or shebangs during dockerfile parsing, these changes only
affect the syntax detection.

Additionally, the relevant functions are moved into the parser package
from the dockerfile2llb package, so that we can share code between the
both the dockerfile directive parsing, and the syntax detection code,
and also removes the dockerfile2llb.ParseDirectives function which
appeared to scan directives in the same was as the parser, but was not
aware of various quirks (e.g. case-insensitivity).

Signed-off-by: Justin Chadwell <me@jedevc.com>