exporter: use local filesystem generation for tar exporter
This connects the tar exporter and the local exporter to the same
fileystem generation, similar to the approach followed by the shared
ImageWriter used in both the containerimage and oci packages.

This allows the tar exporter to use the same attestations generation
logic, so attestations are now supported for the tar exporter.

Signed-off-by: Justin Chadwell <me@jedevc.com>