sbom: fix inconsistencies in sbom protocol
This irons a few issues discovered in porting the default sbom scanner
to go.

Since we construct the args for the image based on the Entrypoint + Cmd,
we shouldn't error out early if no Cmd is set, but only if neither
Entrypoint or Cmd are set.

We should also respect the environment variables set in the config, and
include those, passing them to the ExecOp as well as the custom-set
variables.

We should avoid setting BUILDKIT_SCAN_SOURCE_EXTRAS if no extras have
been specified, to simplify the scanning protocol.

Signed-off-by: Justin Chadwell <me@jedevc.com>