exporter: ensure flat structure with single platform
Previously, the test for generating a nested structures was determined
by the number of refs, which for attestations, can be more than one even
though there is a single platform.

The code is reworked to ensure that if a single platform is specified,
we still get a flat structure, even if attestations are present.

Signed-off-by: Justin Chadwell <me@jedevc.com>