dockerfile: use target name as the dummy reference
Now that the reference is not significant to image resolution, with only
the digest being used to determine what to lookup, the only effect of
the dummy reference is to provide reasonable error messages.

Since the name of the store is likely not significant, using that as the
basis for the reference may be deceiving. We can use the original name
of the image as the dummy reference, except with the targeted digest. We
can't easily require that clients should expose the name of the store,
since that may be a leak of the local path to the store, which may be
private.

Signed-off-by: Justin Chadwell <me@jedevc.com>