exporter: fix single platform fs output with attestations
Previous commit to fix platform output wasn't quite right - while the
original use case for the multi-platform arg works correctly, it was
flawed when used with attestations. We need to fallback to a default of
false if no attestations are present.

multi-platform still works in this case because of a quirk - if a client
has the capability to request attestations, then it will also have the
fix that forwards all FrontendOpts to every section of Build, so we can
extract multi-platform and pass it to the exporter. Not a perfect
solution, but it's not obvious what a more elegant approach would be
without a major refactor.

Signed-off-by: Justin Chadwell <me@jedevc.com>