solver: remove force refs
We no longer need to (or should) force refs conversion.

A single Ref is distinctly different from a Refs map with a single
element - they express the user-specified difference between whether
multi-platform exports should be used or not. It's important to preserve
this sematic, especially on the side of the frontend, where it can't be
recovered if lost.

Signed-off-by: Justin Chadwell <me@jedevc.com>