attestations: remove inline-only metadata key
Clients should instead use the attestations exporter key to control
which exporters should receive which attestations. This allows for
a more fine-grained, client-controlled approach, allowing it to use it's
knowledge of where images end up to determine exactly which attestations
should be allowed.

Additionally, this approach is more flexible, and should hopefully
extend easily to supporting multiple exporters eventually.

Signed-off-by: Justin Chadwell <me@jedevc.com>